Platforms/ARM: Add FileExplorerLib.inf to the dsc file
Dandan submitted this patch to the EDK2 Mailing list for review before
FVP was moved to OpenPlatformPkg:

https://www.mail-archive.com/edk2-devel@lists.01.org/msg07113.html

His original commit message:

ArmPlatformPkg: Add FileExplorerLib.inf to the dsc file

V3: Add FileExplorerLib when SECURE_BOOT_ENABLE == TRUE, so
when to use FileExplorerLib is clear.

Because SecureBootConfigDxe use FileExplorerLib now, but
FileExplorerLib is not in the dsc file of the package
that use SecureBootConfigDxe. Now add it to pass build.

Contributed-under: TianoCore Contribution Agreement 1.0
Signed-off-by: Dandan Bi <dandan.bi@intel.com>
Reviewed-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Acked-by: Ard Biesheuvel <ard.biesheuvel@linaro.org>
Reviewed-by: Leif Lindholm <leif.lindholm@linaro.org>