Platforms/Marvell: Replace ArmDmaLib with NullDmaLib
ArmDmaLib may cause problems for coherent-devices usage. As on
the Edk2 lists:

"ArmDmaLib is fundamentally incompatible with coherent devices. The
code performs a cache invalidate for data written by the bus master,
since on non-coherent systems, this may be necessary to get rid of
stale cachelines that shadow the incoming data. On coherent devices,
the device will write to the cache, not to main memory, and so this
cache invalidation will discard the data coming from the device."

Change-Id: I7cbbfb5d2cb4f9a79162dd7db07269d4bd59b131
Signed-off-by: Marcin Wojtas <mw@semihalf.com>
Reviewed-on: http://vgitil04.il.marvell.com:8080/35124
Tested-by: iSoC Platform CI <ykjenk@marvell.com>
Reviewed-by: Nizan Zorea <nzorea@marvell.com>